requirements.txt
numpy
Pillow
cv2